Use dllexport/dllimport
Something changed in Qt5 signal handling that requires compiling code
with dllimport for the signals to connect.  Otherwise, the result is
a bunch of signal misfirings:

QObject::connect: signal not found in [Class]

In order to export all the symbols correctly so dllimport works,
dllexport also needs to be used, which requires some additional fixes.
So far the resulting binaries appear to be ABI compatible, but
everything needs to be recompiled anyway to use dllimport.